chore(slang): adapt recut to rebased slang dev-solx
Rebasing slang's dev-solx onto origin/main (35 upstream commits) renamed
the linearised-members accessors and changed hex-literal value types:
* `compute_linearised_{functions,state_variables,bases}` →
  `linearised_{functions,state_variables,bases}` (upstream #1805 moved the
  linearisation into its own semantic pass) — renamed all recut call sites.
* `LiteralKind::HexInteger.value` is now an unsigned `BigUint` (was
  `BigInt`); widen it to `BigInt` at the two fold sites.
Re-pinned Cargo.lock to the rebased dev-solx tip. Build + full tester are
byte-identical to before the rebase: 16981 PASSED, same failing set, 0
regressions — purely staying current with upstream slang.

+++ solx-slang/src/ast/contract/getter.rs
@@ -455,7 +455,7 @@ impl<'state, 'context> ContractEmitter<'state, 'context> {
                 .and_then(|slang_type| match slang_type {
                     SlangType::Literal(literal) => match literal.kind() {
                         LiteralKind::Integer { value } => Some(value),
-                        LiteralKind::HexInteger { value, .. } => Some(value),
+                        LiteralKind::HexInteger { value, .. } => Some(BigInt::from(value)),
                         _ => None,
                     },
                     _ => None,
